[#4801] Refactor parse_db_config to avoid importing from old cli
