+++ ckanext/datastore/tests/helpers.py
@@ -3,7 +3,6 @@
 from sqlalchemy import orm
 
 import ckan.model as model
-import ckan.lib.cli as cli
 from ckan.lib import search
 
 import ckan.plugins as p
@@ -41,8 +40,8 @@ def rebuild_all_dbs(Session):
     ''' If the tests are running on the same db, we have to make sure that
     the ckan tables are recrated.
     '''
-    db_read_url_parts = cli.parse_db_config('ckan.datastore.write_url')
-    db_ckan_url_parts = cli.parse_db_config('sqlalchemy.url')
+    db_read_url_parts = model.parse_db_config('ckan.datastore.write_url')
+    db_ckan_url_parts = model.parse_db_config('sqlalchemy.url')
     same_db = db_read_url_parts['db_name'] == db_ckan_url_parts['db_name']
 
     if same_db:
